move admin images out of gameState
