+++ src/admin/store/reducers/imagesReducer.ts
